Simplify photo storage api
