+++ src/photo/PhotoUploadWithStatus.tsx
@@ -9,7 +9,7 @@ import { RefObject, useTransition, useRef, useEffect } from 'react';
 import Spinner from '@/components/Spinner';
 import ResponsiveText from '@/components/primitives/ResponsiveText';
 import { useAppText } from '@/i18n/state/client';
-import { uploadPhotoFromClient } from './storage/client';
+import { uploadPhotoFromClient } from './storage';
 
 export default function PhotoUploadWithStatus({
   inputRef,
