--- src/photo/storage/client.ts
@@ -1,8 +1,0 @@
-import { uploadFileFromClient } from '@/platforms/storage';
-import { EXTENSION_DEFAULT, PREFIX_UPLOAD } from '.';
-
-export const uploadPhotoFromClient = (
-  file: File | Blob,
-  extension = EXTENSION_DEFAULT,
-) =>
-  uploadFileFromClient(file, PREFIX_UPLOAD, extension);
