+++ src/platforms/storage/cache.ts
@@ -2,7 +2,7 @@ import { unstable_noStore } from 'next/cache';
 import {
   getStoragePhotoUrls,
   getStorageUploadUrls,
-} from '@/photo/storage/server';
+} from '@/photo/storage';
 
 export const getStorageUploadUrlsNoStore: typeof getStorageUploadUrls =
   (...args) => {
